Use `java-library` plugin and implementation dependency.
Based on the [wiki
instructions](https://cloudera.atlassian.net/wiki/spaces/ENG/pages/1951170896/Integrate+to+CDP+Structured+Events+WIP#Add-the-module-to-the-build.gradle%3A)
for adding the structured events api, I think the intention was to only
  use the module itself, not needing to include all the deps in every
  service.

I've changed the `build.gradle` for the `structuredevent-service-cdp`
module to use `java-library`, which enables the `api` dependency.

The `api` dependency leaks dependencies from that module in any module
that depends on the `structuredevent-service-cdp` module.

I've updated the Freeipa and Environments build.gradle settings to
remove the extra dependencies.